Added graphics loop to client example

+++ include/ice-engine/memory.h
@@ -72,7 +72,7 @@ namespace ice
 
     };
 
-    template<uint32_t size> class stack_buffer : public std::array<uint8_t,size>
+    template<size_t size> class stack_buffer : public std::array<uint8_t,size>
     {
 
     };
